requirements.txt
appdirs==1.4.3
boto==2.46.1
bz2file==0.98
gensim==2.0.0
gevent==1.2.1
greenlet==0.4.12
numpy==1.12.1
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
smart-open==1.5.2
websocket==0.2.1
xmltodict==0.10.2
